add check time before isolate network; modify log

+++ redis/package/redis-keeper-package/src/main/scripts/upload/upload_files.sh
@@ -12,20 +12,26 @@ if [ -z $local_path ]; then
     exit
 fi
 
+#remote_path="~"
+#mkdir_cmd="mkdir -p $remote_path/network"
 mkdir_cmd="mkdir -p $remote_path"
 
-if [ "$remove" == "rm" ]; then
-    echo "echo rm dir and upload"
-    sshpass -p"$password" ssh -o StrictHostKeyChecking=no $username@$ip "rm -rf $remote_path; $mkdir_cmd" 
-    wait
-else
-    if [ -z $remote_path ]; then
-        echo "echo remote path:~"
-    else
-        echo "echo create remote path: $remote_path"
-        sshpass -p"$password" ssh -o StrictHostKeyChecking=no $username@$ip $mkdir_cmd 
+
+if [ "$remote_path" -ne "~" ] && [ "$remote_path" -ne "/home/xpipe" ]; then
+    if [ "$remove" == "rm" ]; then
+        echo "echo rm dir and upload"
+            sshpass -p"$password" ssh -o StrictHostKeyChecking=no $username@$ip "rm -rf $remote_path; $mkdir_cmd" 
         wait
+    else
+        if [ -z $remote_path ]; then
+            echo "echo remote path:~"
+        else
+            echo "echo create remote path: $remote_path"
+            sshpass -p"$password" ssh -o StrictHostKeyChecking=no $username@$ip $mkdir_cmd 
+            wait
+        fi
     fi
 fi
 sshpass -p"$password" scp -r $local_path $username@$ip:$remote_path
+#sshpass -p"$password" scp -r ~/Downloads/mount.sh $username@$ip:$remote_path
 #sshpass -p"$password" scp -r ../network/* $username@$ip:$remote_path/network
